Enhance save as template and new from template dialogs

See merge request inkscape/inkscape!6937

+++ src/ui/widget/template-list.cpp
@@ -27,6 +27,8 @@
 #include <gtkmm/sortlistmodel.h>
 #include <gtkmm/singleselection.h>
 #include <memory>
+#include <glibmm/miscutils.h>
+#include <gtkmm/filterlistmodel.h>
 
 #include "document.h"
 #include "extension/db.h"
@@ -50,9 +52,10 @@ struct TemplateList::TemplateItem : public Glib::Object {
     Glib::RefPtr<Gdk::Texture> icon;
     Glib::ustring key;
     int priority;
+    Glib::ustring category;
 
     static Glib::RefPtr<TemplateItem> create(const Glib::ustring& name, const Glib::ustring& label, const Glib::ustring& tooltip, 
-        Glib::RefPtr<Gdk::Texture> icon, Glib::ustring key, int priority) {
+        Glib::RefPtr<Gdk::Texture> icon, Glib::ustring key, int priority, const Glib::ustring& category) {
 
         auto item = Glib::make_refptr_for_instance<TemplateItem>(new TemplateItem());
         item->name = name;
@@ -61,6 +64,7 @@ struct TemplateList::TemplateItem : public Glib::Object {
         item->icon = icon;
         item->key = key;
         item->priority = priority;
+        item->category = category;
         return item;
     }
 private:
@@ -69,33 +73,64 @@ private:
 
 
 TemplateList::TemplateList(BaseObjectType *cobject, const Glib::RefPtr<Gtk::Builder> &refGlade)
-    : Gtk::Notebook(cobject)
+    : Gtk::Stack(cobject)
 {
 }
+
+static Glib::ustring all_templates = "All templates";
 
 /**
  * Initialise this template list with categories and icons
  */
-void TemplateList::init(Inkscape::Extension::TemplateShow mode)
+void TemplateList::init(Inkscape::Extension::TemplateShow mode, AddPage add_page, bool allow_unselect)
 {
+    // same width for all items
+    set_hhomogeneous();
+    // height can vary per row
+    set_vhomogeneous(false);
+    // track page switching
+    property_visible_child_name().signal_changed().connect([this]() {
+        _signal_switch_page.emit(get_visible_child_name());
+    });
+
     std::map<std::string, Glib::RefPtr<Gio::ListStore<TemplateItem>>> stores;
 
     Inkscape::Extension::DB::TemplateList extensions;
     Inkscape::Extension::db.get_template_list(extensions);
 
+    Glib::RefPtr<Gio::ListStore<TemplateItem>> all;
+    if (add_page == All) {
+        all = generate_category(all_templates, allow_unselect);
+    }
+
+    int group = 0;
     for (auto tmod : extensions) {
         for (auto preset : tmod->get_presets(mode)) {
             auto const &cat = preset->get_category();
+            if (add_page == Custom && cat != "Custom") continue;
 
             if (auto it = stores.lower_bound(cat);
                 it == stores.end() || it->first != cat)
             {
                 try {
-                    it = stores.emplace_hint(it, cat, generate_category(cat));
+                    group += 10000;
+                    it = stores.emplace_hint(it, cat, generate_category(cat, allow_unselect));
                     it->second->remove_all();
                 } catch (UIBuilderError const& error) {
                     g_error("Error building templates %s\n", error.what());
                     return;
+                }
+
+                if (add_page == Custom) {
+                    // add new template placeholder
+                    auto const filepath = Glib::build_filename("icons", "custom.svg");
+                    auto const fullpath = get_filename(TEMPLATES, filepath.c_str(), false, true);
+                    auto icon = to_texture(icon_to_pixbuf(fullpath, get_scale_factor()));
+                    auto templ = TemplateItem::create(
+                        Glib::Markup::escape_text(_("<new template>")),
+                        "", "", icon, "-new-template-", -1, cat
+                    );
+                    stores[cat]->append(templ);
                 }
             }
 
@@ -106,16 +141,25 @@ void TemplateList::init(Inkscape::Extension::TemplateShow mode)
             auto trans_label = label.empty() ? "" : _(label.c_str());
             auto icon = to_texture(icon_to_pixbuf(preset->get_icon_path(), get_scale_factor()));
 
-            stores[cat]->append(TemplateItem::create(
+            auto templ = TemplateItem::create(
                 Glib::Markup::escape_text(name),
                 Glib::Markup::escape_text(trans_label),
                 Glib::Markup::escape_text(tooltip),
-                icon, preset->get_key(), preset->get_sort_priority()
-            ));
+                icon, preset->get_key(), group + preset->get_sort_priority(),
+                cat
+            );
+            stores[cat]->append(templ);
+            if (all) {
+                all->append(templ);
+            }
         }
     }
 
-    reset_selection();
+    refilter(_search_term);
+
+    if (allow_unselect) {
+        reset_selection();
+    }
 }
 
 /**
@@ -133,45 +177,60 @@ Cairo::RefPtr<Cairo::ImageSurface> TemplateList::icon_to_pixbuf(std::string cons
         return cache[path];
     }
     Inkscape::svg_renderer renderer(path.c_str());
-    auto result = renderer.render_surface(scale);
+    auto result = renderer.render_surface(scale * 0.7); // reduced template icon size to fit more in a dialog
     cache[path] = result;
     return result;
+}
+
+sigc::signal<void (const Glib::ustring&)> TemplateList::signal_switch_page() {
+    return _signal_switch_page;
 }
 
 /**
  * Generate a new category with the given label and return it's list store.
  */
-Glib::RefPtr<Gio::ListStore<TemplateList::TemplateItem>> TemplateList::generate_category(std::string const &label)
+Glib::RefPtr<Gio::ListStore<TemplateList::TemplateItem>> TemplateList::generate_category(std::string const &label, bool allow_unselect)
 {
     auto builder = create_builder("widget-new-from-template.ui");
-    auto container = &get_widget<Gtk::ScrolledWindow>(builder, "container");
-    auto icons     = &get_widget<Gtk::GridView>      (builder, "iconview");
+    auto& container = get_widget<Gtk::ScrolledWindow>(builder, "container");
+    auto& icons     = get_widget<Gtk::GridView>      (builder, "iconview");
 
     auto store = Gio::ListStore<TemplateItem>::create();
     auto sorter = Gtk::NumericSorter<int>::create(Gtk::ClosureExpression<int>::create([this](auto& item){
-        auto ptr = std::dynamic_pointer_cast<TemplateList::TemplateItem>(item);
+        auto ptr = std::dynamic_pointer_cast<TemplateItem>(item);
         return ptr ? ptr->priority : 0;
     }));
-    auto model = Gtk::SortListModel::create(store, sorter);
-    auto selection_model = Gtk::SingleSelection::create(model);
-    selection_model->set_can_unselect();
+    auto sorted_model = Gtk::SortListModel::create(store, sorter);
+    if (!_filter) {
+        _filter = Gtk::BoolFilter::create({});
+    }
+    auto filtered_model = Gtk::FilterListModel::create(sorted_model, _filter);
+    auto selection_model = Gtk::SingleSelection::create(filtered_model);
+    if (allow_unselect) {
+        selection_model->set_can_unselect();
+        selection_model->set_autoselect(false);
+    }
     auto factory = IconViewItemFactory::create([](auto& ptr) -> IconViewItemFactory::ItemData {
         auto tmpl = std::dynamic_pointer_cast<TemplateItem>(ptr);
         if (!tmpl) return {};
 
-        auto label = tmpl->name + "\n<small><span alpha=\"60%\" line_height=\"1.75\">" + tmpl->label + "</span></small>";
+        auto label = tmpl->label.empty() ? tmpl->name :
+            tmpl->name + "<small><span line_height='0.5'>\n\n</span><span alpha='60%'>" + tmpl->label + "</span></small>";
         return { .label_markup = label, .image = tmpl->icon, .tooltip = tmpl->tooltip };
     });
-    icons->set_factory(factory->get_factory());
-    icons->set_model(selection_model);
+    icons.set_max_columns(30);
+    icons.set_tab_behavior(Gtk::ListTabBehavior::ITEM); // don't steal the tab key
+    icons.set_factory(factory->get_factory());
+    icons.set_model(selection_model);
 
     // This packing keeps the Gtk widget alive, beyond the builder's lifetime
-    append_page(*container, g_dpgettext2(nullptr, "TemplateCategory", label.c_str()));
+    add(container, label, g_dpgettext2(nullptr, "TemplateCategory", label.c_str()));
+    _categories.emplace_back(label);
 
     selection_model->signal_selection_changed().connect([this](auto pos, auto count){
-        _item_selected_signal.emit();
+        _item_selected_signal.emit(count > 0 ? static_cast<int>(pos) : -1);
     });
-    icons->signal_activate().connect([this](auto pos){
+    icons.signal_activate().connect([this](auto pos){
         _item_activated_signal.emit();
     });
 
@@ -184,20 +243,34 @@ Glib::RefPtr<Gio::ListStore<TemplateList::TemplateItem>> TemplateList::generate_
  */
 bool TemplateList::has_selected_preset()
 {
-    return (bool)get_selected_preset();
+    return !!get_selected_preset();
+}
+
+bool TemplateList::has_selected_new_template() {
+    if (auto item = get_selected_item()) {
+        return item->key == "-new-template-";
+    }
+    return false;
+}
+
+Glib::RefPtr<TemplateList::TemplateItem> TemplateList::get_selected_item(Gtk::Widget* current_page) {
+    if (auto iconview = get_iconview(current_page ? current_page : get_visible_child())) {
+        auto sel = std::dynamic_pointer_cast<Gtk::SingleSelection>(iconview->get_model());
+        auto ptr = sel->get_selected_item();
+        if (auto item = std::dynamic_pointer_cast<TemplateList::TemplateItem>(ptr)) {
+            return item;
+        }
+    }
+    return nullptr;
 }
 
 /**
  * Returns the selected template preset, if one is not selected returns nullptr.
  */
-std::shared_ptr<TemplatePreset> TemplateList::get_selected_preset()
+std::shared_ptr<TemplatePreset> TemplateList::get_selected_preset(Gtk::Widget* current_page)
 {
-    if (auto iconview = get_iconview(get_nth_page(get_current_page()))) {
-        auto sel = std::dynamic_pointer_cast<Gtk::SingleSelection>(iconview->get_model());
-        auto ptr = sel->get_selected_item();
-        if (auto item = std::dynamic_pointer_cast<TemplateList::TemplateItem>(ptr)) {
-            return Extension::Template::get_any_preset(item->key);
-        }
+    if (auto item = get_selected_item(current_page)) {
+        return Extension::Template::get_any_preset(item->key);
     }
     return nullptr;
 }
@@ -205,10 +278,10 @@ std::shared_ptr<TemplatePreset> TemplateList::get_selected_preset()
 /**
  * Create a new document based on the selected item and return.
  */
-SPDocument *TemplateList::new_document()
+SPDocument *TemplateList::new_document(Gtk::Widget* current_page)
 {
     auto app = InkscapeApplication::instance();
-    if (auto preset = get_selected_preset()) {
+    if (auto preset = get_selected_preset(current_page)) {
         if (auto doc = preset->new_from_template()) {
             // TODO: Add memory to remember this preset for next time.
             return app->document_add(std::move(doc));
@@ -221,13 +294,55 @@ SPDocument *TemplateList::new_document()
     return app->document_new();
 }
 
+// Show page by its name
+void TemplateList::show_page(const Glib::ustring& name) {
+    set_visible_child(name);
+    refilter(_search_term);
+}
+
+// callback to check if template should be visible
+bool TemplateList::is_item_visible(const Glib::RefPtr<Glib::ObjectBase>& item, const Glib::ustring& search) const {
+    auto ptr = std::dynamic_pointer_cast<TemplateItem>(item);
+    if (!ptr) return false;
+
+    const auto& templ = *ptr;
+
+    if (search.empty()) return true;
+
+    // filter by name and label
+    return templ.label.lowercase().find(search) != Glib::ustring::npos ||
+           templ.name.lowercase().find(search) != Glib::ustring::npos;
+}
+
+// filter list of visible templates down to those that contain given search string in their name or label
+void TemplateList::filter(Glib::ustring search) {
+    _search_term = search;
+    refilter(search);
+}
+
+// set keyboard focus on template list
+void TemplateList::focus() {
+    if (auto iconview = get_iconview(get_visible_child())) {
+        iconview->grab_focus();
+    }
+}
+
+void TemplateList::refilter(Glib::ustring search) {
+    // When a new expression is set in the BoolFilter, it emits signal_changed(),
+    // and the FilterListModel re-evaluates the filter.
+    search = search.lowercase();
+    auto expression = Gtk::ClosureExpression<bool>::create([this, search](auto& item){ return is_item_visible(item, search); });
+    // filter results
+    _filter->set_expression(expression);
+}
+
 /**
  * Reset the selection, forcing the use of the default template.
  */
-void TemplateList::reset_selection()
+void TemplateList::reset_selection(Gtk::Widget* current_page)
 {
     // TODO: Add memory here for the new document default (see new_document).
-    for (auto const widget : UI::get_children(*this)) {
+    for (auto const widget : UI::get_children(current_page ? *current_page : *this)) {
         if (auto iconview = get_iconview(widget)) {
             auto sel = std::dynamic_pointer_cast<Gtk::SingleSelection>(iconview->get_model());
             sel->unselect_all();
